Conflicts:
	ClntTransMgr/ClntTransMgr.cpp
	Misc/DHCPConst.h
	Requestor/ReqTransMgr.h
	Requestor/Requestor.cpp
	SrvIfaceMgr/SrvIfaceMgr.cpp
	SrvMessages/SrvMsg.cpp
	SrvMessages/SrvMsgLeaseQueryReply.cpp
	SrvTransMgr/SrvTransMgr.cpp

+++ Options/OptUserClass.h
@@ -18,8 +18,7 @@ class TOptUserClass : public TOpt
 {
  public:
     TOptUserClass( char * &buf,  int &n, TMsg* parent);
-    
-    int getSize();
+    size_t getSize();
     bool isValid();
     char * storeSelf( char* buf);
 };
